Implement strain boundary conditions for tension, compression, shearing (only along the three cartesian directions) and torsion (only twisting at the top).
Changes were made to the run_strain_sim() function and get_accel_scaled(). Future changes to get_accel_scaled() may be necessary to allow for the case of compression where transverse motion of the nodes is allowed. For now, all nodes invovled in a boundary that the strain is applied to are held fixed.