Implement extend_simulate_scaled() in simulate module and extend_from_checkpoint() in checkpoint_analysis.py, for continuing or extending simulations that were run with persistent checkpointing. Implement (incomplete) tension and compression effective modulus calculation as a function of applied strain.
extending simulations from checkpoints can be done for checkpoint files that did not come from persistent checkpointing, but that requires separate implementation, as the state (applied strain, field, etc) of the checkpoint file must be inferred in a manner different from the persistent checkpointing files. tension/compression modulus calculation is a subfunction of get_effective_modulus(), which is currently in polymer_tension_analysis.py. The functionality should cover tension, compression, shearing, and torsion, but it will also need to be generalized, or a separate functionality implemented, for calculating effective moduli as a function of applied field