Implement calculation and plotting of effective stress and effective modulus versus checkpoint number by analyzing (persistent checkpointing) checkpoint files.
Each checkpoint file represents the state of the simulated system at a different point in the simulation/integration. By studying the behavior of the effective stress versus checkpoint number we are able to see if the values converge or not, and develop a better understanding of the overall convergence of the system. This should assist in making better choices for the number of integrations within an integration step, and the number of integration steps allowed. While the nodes may still be undergoing some rearrangement, for the purposes of studying the effective material properties the exact final arrangement may be less important.