Implement updated gpu calculation of spring and volume correction forces. Implement calculation, plotting, and comparison of magnetization and magnetic force calculations in magnetism_testing.py
gpu calculations modified by including l_e (discretization length) scaling in calculation of kappa and k stiffness constants, allowing a singular force array to be instantiated on the gpu and passed back to the cpu, ideally cutting down on data transfer times. additionally, scaling by beta_i may be implemented as a kernel and done on the gpu in the next commit, after profiling has been performed to see the impact of the current set of changes.