Implement a second gpu based approach to simulating the MRE system. Implement a leapfrog type integrator using cupy RawKernel. Implement drag force calculation and beta_i scaling using cupy RawKernels.
Still in the process of debugging. Currently trying to understand the presence of a strong force in the y direction for the case of an attractive applied magnetic field configuration. Assumed to be part of the composite force calculation function that is handling the calculation of VCF, spring, and drag forces, and the scaling by beta_i.