Implement changes to 32bit magnetic dipole-dipole interaction force calculations in the magnetism module, such that the particle positions are given in SI units (meters) and the returned force is in SI units (Netwons)
This is an attempt to deal with errors that can arise due to the use of 32bit floating point representation and the very small values returned when normalized positions were used to return a force that was then scaled by beta after the function call. While initial results are promising, allowing integration to run for more steps (3000 vs 1000 per integration round, and 3 rounds of integration), by the end of the second integration round the particles began separating further from each other (at 0.1T with non-zero compressive stress applied) and the analysis of acceleration norms results in NaN values. something is breaking and I'm not sure why, or when.