Implement fix to get_nodes_from_grid_voxels() used for particle placement.
Translation of the particle node positions based on the desired center was not performing as desired. Due to details of the implementation, the argument grid_points was an array of arrays of 3 integer values, strictly positive. as a consequence, instead of the center of the sphere being translated to the desired point, the [0,0,0] point from the grid_points argument was translated to that position. Increased voxel radius of the spheres would result in greater deviations of particle placement from the desired placement.