Update implementation of magnetization finding to handle zero field cases gracefully. Update simulate_scaled_gpu_leapfrog_v3() to not consider particle nodes when checkiong convergence criteria. Implement gpu based particle position finding to remove memory transfer overhead.
By only considering polymer nodes, the motion of the particles is still captured indirectly through the deformation of the polymer. Rigidity of the particle-particle springs results in strong accelerations and velocities of the nodes(as they oscillate about equilibrium separation) despite overall particle translation and rotation having ended. GPU particle position finding required changes to sphere_rasterization.py to set the first 8 elements of each row of the particles variable to be the vertices of the central voxel of the magnetic particle.